+++ src/main/scala/com/nawforce/rpc/Server.scala
@@ -129,6 +129,10 @@ class Server(child: ChildProcess) {
     orgAPI.getDefinition(path, line, offset, content)
   }
 
+  def getImplementation(path: String, line: Int, offset: Int, content: Option[String]): Future[Array[LocationLink]] = {
+    orgAPI.getImplementation(path, line, offset, content)
+  }
+
   def getDependencyBombs(count: Int): Future[Array[BombScore]] = {
     orgAPI.getDependencyBombs(count)
   }